extension-py/requirements.txt
websocket-client==1.8.0
